move validators to separate file so they are omitted in generated doc output
