+++ scripts/tdoptions.json
@@ -6,10 +6,12 @@
   "mode": "file",
   "out": "docs",
   "src": [
-    "src/index.ts",
+    "typings/es6-collections/es6-collections.d.ts",
     "typings/es6-promise/es6-promise.d.ts",
     "typings/text-encoding/text-encoding.d.ts",
-    "typings/es6-container-shim.d.ts",
-    "typings/es6-dataview.d.ts"
+    "typings/es6-dataview.d.ts",
+    "node_modules/phosphor-signaling/lib/phosphor-signaling.d.ts",
+    "src/kernel.ts",
+    "src/session.ts"
   ]
 }
